fix: updated controller flag parsing to support slices of arguments
Before this change was created when a flag that contained a slice of
arguments was processed it would show up in the command like --flag
"[value1 value2]". The command would see this as a single value and
process it incorrectly. The same would occur with a slice of one value
for example, --flag "[value1]". After this change the flag is split into
multiple flags, --flag "value1" --flag "value2". I figured this would be
safer than trying to escape commas and concat the values around a comma.