feat: allow tools to be generated from any cobra.Command (#13)
Before this change it was always assumed that ophis would be mounted as
a root command. This means that if ophis were to be mounted under a
subcommand like "alpha" that ophis would only scan the alpha subcommand
for tools. With this change the user can pass in any root command and
have ophis scan that command for tools.

Claude and VSCode mcp subcommand also correctly use the path that the
MCP server was mounted at.

* feat: allow tools to be generated from any cobra.Command

Before this change it was always assumed that ophis would be mounted as
a root command. This means that if ophis were to be mounted under a
subcommand like "alpha" that ophis would only scan the alpha subcommand
for tools. With this change the user can pass in any root command and
have ophis scan that command for tools.

Claude and VSCode mcp subcommand also correctly use the path that the
MCP server was mounted at.